Updated nmap to python
Updated scanning ports from nmap to python with zeroconf to increase sppeed.

+++ persist.sh
@@ -1,48 +1,59 @@
 echo kabhooom!
-
+# set -x
 # prerequisites(){
     
 
 _confirm() {
+    local l="${2:-3}"
   echo "$1, if yes, then continue."
   echo
-  read -p "Continue? [Y/n]: " answer
-  answer="${answer:-y}"
-  if [ "$answer" != "y" ] && [ "$answer" != "Y" ]; then
-    # echo "$2"
+  read -p "Continue? [Y/n]: " a
+  local a="${a:-y}"
+  if [ "$a" != "y" ] && [ "$a" != "Y" ]; then
+    # echo "Aborted."
     
-    printf '\033[3A\033[J'
+    printf "\033[${l}A\033[J"
     return 1
   fi
-  printf '\033[3A\033[J'
+  printf "\033[${l}A\033[J"
 }
 # _confirm "Turned on dev options?"
 # exit 0;
 
 _usb_debug_ack() {
-  getprop init.svc.adbd
+    getprop init.svc.adbd
 }
 
-_get_devices(){
-    adb devices 2>/dev/null | grep -A 100 'List of devices attached' | tail -n +2 # | cut -sf -1
+_get_devices() {
+  local result=$(adb devices 2>/dev/null | grep -A 100 'List of devices attached' | tail -n +2)
+  if [ -z "$result" ]; then
+    return 1
+  fi
+  echo "$result"
 }
 _set_port(){
     if [ "$(_usb_debug_ack)" != "running" ]; then
         echo "USB debugging is disabled. Turn on the USB debugging and wireless debugging to set the port."
         echo "and make sure the device is paired."
-        echo "launching developer options..."
-        sleep 2;
+        echo
+        _confirm "About to launch developer options..." 7 || exit 1;
+        echo "Launching developer options..."
+        # sleep 2
         am start -n "com.android.settings/.Settings\$DevelopmentSettingsDashboardActivity" >/dev/null 2>&1
         exit 1
     fi
-    PORT=$(nmap -sT -p30000-60000 --open localhost | grep '^ *[0-9]' | cut -d/ -f1 || echo "port not found")
+    # nmap -sT -p30000-60000 --open localhost | grep '^ *[0-9]' | cut -d/ -f1
+    # python3 ~/get_adbWifi_port.py 2>/dev/null
+    PORT=$(python3 get_adbWifi_port.py 2>/dev/null)
     if [[ -z "$PORT" ]]; then
         echo "Couldn't find port: $PORT"
         echo "make sure to turn on the wireless debugging and pair the device."
-        exit 1;
+        exit 1
     fi
-    echo adb port: $PORT;
+    echo "ADB port: $PORT"
 }
+# _set_port
+# exit 0;
 _set_ip() {
     for interface in wlan0 v4-ccmni1 v4-ccmni2; do
         local candidate=$(ifconfig 2>/dev/null | grep -A 3 "$interface" | grep 'inet ' | awk '{print $2}')
@@ -69,6 +80,37 @@ _adb_connection_ack() {
     fi
     adb connect "$IP:$PORT" # | tee /dev/stderr
 }
+persist_adb(){
+    
+    echo "hdjjs"
+    
+    echo "printing devices from persistence..."
+    _get_devices || { echo "can't print devices"; kill $(pgrep -x adb); return 1; }
+    # (_get_devices || echo "can't print devices") && return 1;
+    # sleep 10
+    echo "starting adb tunnel as tcp..."
+    result=$(adb tcpip 5555)
+    if ! echo "$result" | grep -q "TCP"; then
+      echo "can't connect to tcp ip: $result"
+      return 1
+    fi
+    echo
+    # sleep 2
+    echo "killing old adb sessions..."
+    adb kill-server 2>/dev/null;
+    echo
+    # sleep 2
+    echo "starting fresh adb server..."
+    adb start-server 2>/dev/null
+    # sleep 2
+    echo "turning off wireless debugging."
+    adb shell settings put global adb_wifi_enabled 0;
+    echo
+    echo "printing devices: "
+    _get_devices 
+    exit 0;
+}
+
 _post_pairing_notification(){
     local REPLY_FILE="/data/data/com.termux/files/home/.tmp/reply.txt"
     rm -f "$REPLY_FILE"     # clean previous if exists
@@ -127,6 +169,10 @@ pair(){
 }
 stop(){
     echo "Stopping..."
+    if ! pgrep -x adb ; then
+        echo "no adb process found, exiting."
+        exit 0;
+    fi
     if [ -z "$(_get_devices)" ]; then
         echo "No devices attached."
         kill -9 "$(pgrep -x adb)" >/dev/null 2>&1
@@ -153,15 +199,23 @@ stop(){
     adb kill-server
     kill -9 "$(pgrep -x adb)" >/dev/null 2>&1
     
+    echo "pgrep -x adb: $(pgrep -x adb)"
+    echo "pgrep -x socat: $(pgrep -x socat)"
+    
 }
 start(){
     clear
     echo "checking whether the adb is already connected..."
+    
     if [ -n "$(pgrep -x adb)" ]; then
-        if [ -n "$(_get_devices)" ]; then
-            echo "already connected"
+        echo "adb process found."
+        if _get_devices | grep -q device; then 
+            echo "found connected devices."
             _get_devices
-            exit 0;
+            if ! _get_devices | grep -q '^emulator'; then
+                echo "but no emulators found."
+                persist_adb
+            fi
         fi
     fi
     echo "adb is not connected."
@@ -170,15 +224,29 @@ start(){
     if ! _confirm "Do you turned on developer options?"; then 
         echo "Turn on the developer options."
         echo "Tap on (build number/OS version) number 7 times."
-        _confirm "About to launch \"About phone\" settings" || exit 1;
+        echo
+        _confirm "About to launch \"About phone\" settings" 6 || exit 1;
         echo "Launching settings..."
         am start -a android.settings.DEVICE_INFO_SETTINGS  >/dev/null 2>&1
         exit 1;
 
     fi
+    if [ "$(_usb_debug_ack)" != "running" ]; then
+        echo "USB debugging isn't turned on."
+        _confirm "Want me to take to dev screen?" 4
+        echo "Launching dev screen..."
+        
+        am start -n "com.android.settings/.Settings\$DevelopmentSettingsDashboardActivity" >/dev/null 2>&1
+        
+        
+        exit 1;
+    fi
+    
     if ! _confirm "Do you turned on wireless debugging?" ; then 
         echo "Turn on the wireless debugging."
-        echo "Launching wireless debugging settings."
+        echo
+        _confirm "About to launch wireless debugging settings." 6 || exit 1;
+        echo "Launching wireless debugging..."
         am start -n com.android.settings/.SubSettings \
             -e :android:show_fragment com.android.settings.development.WirelessDebuggingFragment \
             -e :android:show_fragment_title "Wireless Debugging" >/dev/null 2>&1
@@ -217,24 +285,8 @@ start(){
     
     echo "adb connection established successfully.";
     sleep 2
-    echo "printing devices..."
-    adb devices
-    sleep 2
-    echo "starting adb tunnel as tcp..."
-    adb tcpip 5555;
-    echo
-    sleep 2
-    echo "killing old adb sessions..."
-    adb kill-server 2>/dev/null;
-    echo
-    sleep 2
-    echo "starting fresh adb server..."
-    adb start-server 2>/dev/null
-    sleep 2
-    echo "turning off wireless debugging."
-    adb shell settings put global adb_wifi_enabled 0;
-    echo
-    adb devices 
+    
+    persist_adb
     
 }
 status(){
